requirements-test_batch_release_charms.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements-test_batch_release_charms.in
#
anyio==3.7.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   httpcore
attrs==23.1.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   jsonschema
certifi==2023.7.22
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   cryptography
    #   pynacl
charmed-kubeflow-chisme==0.2.0
    # via -r ./scripts/release-charms/requirements.txt
charset-normalizer==3.2.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   requests
click==8.1.6
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   typer
cryptography==41.0.3
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pyjwt
deepdiff==6.2.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
deprecated==1.2.14
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pygithub
exceptiongroup==1.1.2
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   anyio
    #   pytest
h11==0.14.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   httpcore
httpcore==0.17.3
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   httpx
httpx==0.24.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   lightkube
idna==3.4
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.0.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   jsonschema
iniconfig==2.0.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pytest
jinja2==3.1.2
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
jsonschema==4.17.3
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   serialized-data-interface
lightkube==0.14.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
lightkube-models==1.27.1.4
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   lightkube
markupsafe==2.1.3
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   jinja2
ops==2.5.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   deepdiff
packaging==23.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pytest
pkgutil-resolve-name==1.3.10
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   jsonschema
pluggy==1.2.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pytest
pycparser==2.21
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   cffi
pygithub==1.59.1
    # via -r ./scripts/release-charms/requirements.txt
pyjwt[crypto]==2.8.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pygithub
pynacl==1.5.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pygithub
pyrsistent==0.19.3
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   jsonschema
pytest==7.4.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   -r requirements-test_batch_release_charms.in
    #   pytest-mock
pytest-mock==3.11.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   -r requirements-test_batch_release_charms.in
pyyaml==6.0.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pygithub
    #   serialized-data-interface
ruamel-yaml==0.17.32
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   ruamel-yaml
serialized-data-interface==0.7.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
sniffio==1.3.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.2.2
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   charmed-kubeflow-chisme
tomli==2.0.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   pytest
typer==0.9.0
    # via -r ./scripts/release-charms/requirements.txt
typing-extensions==4.7.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   typer
urllib3==2.0.4
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   requests
websocket-client==1.6.1
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   ops
wrapt==1.15.0
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   deprecated
zipp==3.16.2
    # via
    #   -r ./scripts/release-charms/requirements.txt
    #   importlib-resources
